+++ src/models/Settings.php
@@ -10,4 +10,7 @@ class Settings extends Model
     public $enabled = false;
     public $category = 'audit';
     public $level = Logger::LEVEL_INFO;
+    public $frontend = false;
+    public $backend = true;
+    public $properties = false;
 }